requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.22
pdf2image==1.17.0

# OCR & Text Extraction
pytesseract==0.3.10
pillow<10.0.0  # Required for EasyOCR compatibility (ANTIALIAS support)
opencv-python==4.9.0.80
easyocr==1.7.0  # Free OCR for figure text extraction
numpy<2.0.0  # Required for compatibility with EasyOCR and other packages

# Machine Learning & Deep Learning
torch==2.2.0
torchvision==0.17.0
transformers==4.38.1
timm==1.0.19

# Table Processing
pandas==2.2.0
tabulate==0.9.0
openpyxl==3.1.2

# Image Processing
scikit-image==0.22.0
matplotlib==3.8.3

# Document Processing
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
pygments==2.19.2

# API Clients
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3

# Math & LaTeX
sympy==1.12
latex2sympy2==1.9.1

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
colorama==0.4.6
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.40.6

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.2.0
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2

# Cloud Storage
b2sdk==1.21.0  # Backblaze B2 SDK

# Message Queue
pika==1.3.2  # RabbitMQ client (legacy)
aio-pika==9.3.1  # Async RabbitMQ client

# External Service Dependencies
# Note: GROBID, PDFFigures2, and Nougat run as separate services
redis==5.0.1  # For caching
celery==5.3.4  # For async processing